+++ tests/testthat/test-verb-select.R
@@ -374,6 +374,13 @@ test_that("mutate preserves grouping vars (#396)", {
   expect_equal(df %>% mutate(b = 1) %>% op_grps(), c("a", "b"))
 })
 
+test_that("select after arrange(desc()) works (#1206)", {
+  out <- lazy_frame(x = 1, y = 1) %>%
+    arrange(desc(x)) %>%
+    select(x)
+  expect_equal(op_vars(out), "x")
+})
+
 
 # lazy_select_query -------------------------------------------------------